update 02 factory method

+++ 02_factory-method_zahlen-strings/src/IntFactory.java
@@ -2,6 +2,6 @@ public class IntFactory extends Factory {
 
     @Override
     public Number getNumber(String s) throws NumberFormatException {
-        return Integer.parseInt(s);
+        return (int) Double.parseDouble(s);
     }
 }
