+++ 02_factory-method_zahlen-strings/src/Main.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
 
         try {
-            Factory factory = Factory.getInstance(FactoryType.DOUBLE);
+            Factory factory = Factory.getInstance(FactoryType.INT);
 
             System.out.println(factory.getNumber("111.3"));
